requirements.txt
snowflake-connector-python>=3.0.0
streamlit>=1.30.0
